+++ packages/@expo/fingerprint/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 🐛 Bug fixes
 
 - Update to `getenv@2.0.0` to support upper case boolean environment variables ([#36688](https://github.com/expo/expo/pull/36688) by [@stephenlacy](https://github.com/stephenlacy))
+- Added `--transitive-linking-dependencies` for `react-native-edge-to-edge` autolinking. ([#37209](https://github.com/expo/expo/pull/37209) by [@kudo](https://github.com/kudo))
 
 ### 💡 Others